@@ -1,6 +1,6 @@
 {
   "name": "dynamic-mapping",
-  "version": "4.5.1",
+  "version": "4.6.0",
   "author": "christof.strack@softwareag.com, stefan.witschel@softwareag.com",
   "description": "Cumulocity plugin to map custom JSON payloads to C8Y payloads.The plugin support both directions: inbound/outbound. Currently MQTT is supported ",
   "repository": {
@@ -9,7 +9,7 @@
   },
   "scripts": {
     "start": "ng serve",
-    "start:admin": "npm start -- --shell administration",
+    "start:admin": "ng serve -- --shell administration",
     "build": "ng build",
     "deploy": "ng deploy",
     "format": "prettier --write 'src/**/*.ts'",
@@ -17,44 +17,47 @@
     "cy:run": "cypress run"
   },
   "keywords": [],
-  "license": "ISC",
+  "license": "Apache-2.0",
   "dependencies": {
-    "@angular/animations": "^17.3.0",
-    "@angular/common": "^17.3.0",
-    "@angular/compiler": "^17.3.0",
-    "@angular/core": "^17.3.0",
-    "@angular/forms": "^17.3.0",
-    "@angular/platform-browser": "^17.3.0",
-    "@angular/platform-browser-dynamic": "^17.3.0",
-    "@angular/platform-server": "^17.3.0",
-    "@angular/router": "^17.3.0",
-    "@angular/upgrade": "^17.3.0",
-    "@c8y/bootstrap": "^1020.18.1",
-    "@c8y/ngx-components": "^1020.18.1",
-    "@c8y/options": "^1020.18.1",
+    "@angular/animations": "^18.0.0",
+    "@angular/common": "^18.0.0",
+    "@angular/compiler": "^18.0.0",
+    "@angular/core": "^18.0.0",
+    "@angular/forms": "^18.0.0",
+    "@angular/platform-browser": "^18.0.0",
+    "@angular/platform-browser-dynamic": "^18.0.0",
+    "@angular/platform-server": "^18.0.0",
+    "@angular/router": "^18.0.0",
+    "@angular/upgrade": "^18.0.0",
+    "@c8y/bootstrap": "1021.0.11",
+    "@c8y/client": "1021.0.11",
+    "@c8y/devkit": "1021.0.11",
+    "@c8y/ngx-components": "1021.0.11",
+    "@c8y/options": "1021.0.11",
+    "@c8y/package-blueprint": "1021.0.11",
+    "@c8y/style": "1021.0.11",
     "@ngx-translate/core": "15.0.0",
     "css-loader": "^7.1.2",
     "echarts": "^5.5.1",
     "jsonata": "^2.0.3",
-    "ngx-bootstrap": "^12.0.0",
+    "ngx-bootstrap": "^18.0.0",
     "ngx-echarts": "^18.0.0",
     "postcss-loader": "^8.1.0",
     "rxjs": "~7.8.1",
     "style-loader": "^4.0.0",
-    "vanilla-jsoneditor": "^0.23.0",
+    "vanilla-jsoneditor": "^2.0.0",
+    "webpack": "5.90.3",
     "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.8",
+    "@angular-devkit/build-angular": "^18.2.11",
     "@angular-eslint/eslint-plugin": "^18.1.0",
     "@angular-eslint/eslint-plugin-template": "^18.3.0",
-    "@angular-eslint/template-parser": "^17.1.1",
-    "@angular/compiler-cli": "^17.3.11",
-    "@angular/language-service": "^18.2.1",
-    "@angular/localize": "^17.3.0",
-    "@angular/service-worker": "^17.3.0",
-    "@c8y/devkit": "^1020.18.1",
-    "@c8y/package-blueprint": "^1020.18.1",
+    "@angular-eslint/template-parser": "^18.3.0",
+    "@angular/compiler-cli": "^18.2.10",
+    "@angular/language-service": "^18.2.2",
+    "@angular/localize": "^18.0.0",
+    "@angular/service-worker": "^18.0.0",
     "@typescript-eslint/eslint-plugin": "^7.17.0",
     "cumulocity-cypress": "^0.4.6",
     "cypress": "^13.13.3",
