@@ -98,10 +98,17 @@ public class App {
         executor.setQueueCapacity(25);
         return executor;
     }
+    //Assuming we can process 25 messages in parallel per CPU-Core
+    @Bean("processingCachePool")
+    public ExecutorService processingThreadPool() {
+        return Executors.newCachedThreadPool();
+        //return Executors.newFixedThreadPool(Runtime.getRuntime().availableProcessors()*25);
+    }
 
+    //Assuming we can process 10 messages in parallel per CPU-Core
     @Bean("cachedThreadPool")
     public ExecutorService cachedThreadPool() {
-        return Executors.newCachedThreadPool();
+        return Executors.newFixedThreadPool(Runtime.getRuntime().availableProcessors()*10);
     }
 
     @Bean
