@@ -34,51 +34,61 @@ import lombok.ToString;
 @ToString()
 @AllArgsConstructor
 public class ServiceConfiguration implements Cloneable {
-    public ServiceConfiguration() {
-        this.logPayload = false;
-        this.logSubstitution = false;
-        this.logConnectorErrorInBackend = false;
-        this.sendConnectorLifecycle = false;
-        this.sendMappingStatus = true;
-        this.sendSubscriptionEvents = false;
-        this.sendNotificationLifecycle = false;
-        this.externalExtensionEnabled = true;
-        this.outboundMappingEnabled = true;
-    }
+	public ServiceConfiguration() {
+		this.logPayload = false;
+		this.logSubstitution = false;
+		this.logConnectorErrorInBackend = false;
+		this.sendConnectorLifecycle = false;
+		this.sendMappingStatus = true;
+		this.sendSubscriptionEvents = false;
+		this.sendNotificationLifecycle = false;
+		this.externalExtensionEnabled = true;
+		this.outboundMappingEnabled = true;
+		this.inboundExternalIdCacheSize = 0;
+		this.inboundExternalIdCacheRetention = 1;
+	}
 
-    @NotNull
-    @JsonSetter(nulls = Nulls.SKIP)
-    public boolean logPayload;
+	@NotNull
+	@JsonSetter(nulls = Nulls.SKIP)
+	public boolean logPayload;
 
-    @NotNull
-    @JsonSetter(nulls = Nulls.SKIP)
-    public boolean logSubstitution;
+	@NotNull
+	@JsonSetter(nulls = Nulls.SKIP)
+	public boolean logSubstitution;
 
-    @NotNull
-    @JsonSetter(nulls = Nulls.SKIP)
-    public boolean logConnectorErrorInBackend;
+	@NotNull
+	@JsonSetter(nulls = Nulls.SKIP)
+	public boolean logConnectorErrorInBackend;
 
-    @NotNull
-    @JsonSetter(nulls = Nulls.SKIP)
-    public boolean sendConnectorLifecycle;
+	@NotNull
+	@JsonSetter(nulls = Nulls.SKIP)
+	public boolean sendConnectorLifecycle;
 
-    @NotNull
-    @JsonSetter(nulls = Nulls.SKIP)
-    public boolean sendMappingStatus;
+	@NotNull
+	@JsonSetter(nulls = Nulls.SKIP)
+	public boolean sendMappingStatus;
 
-    @NotNull
-    @JsonSetter(nulls = Nulls.SKIP)
-    public boolean sendSubscriptionEvents;
+	@NotNull
+	@JsonSetter(nulls = Nulls.SKIP)
+	public boolean sendSubscriptionEvents;
 
-    @NotNull
-    @JsonSetter(nulls = Nulls.SKIP)
-    public boolean sendNotificationLifecycle;
+	@NotNull
+	@JsonSetter(nulls = Nulls.SKIP)
+	public boolean sendNotificationLifecycle;
 
-    @NotNull
-    @JsonSetter(nulls = Nulls.SKIP)
-    public boolean externalExtensionEnabled;
+	@NotNull
+	@JsonSetter(nulls = Nulls.SKIP)
+	public boolean externalExtensionEnabled;
 
-    @NotNull
-    @JsonSetter(nulls = Nulls.SKIP)
-    public boolean outboundMappingEnabled;
+	@NotNull
+	@JsonSetter(nulls = Nulls.SKIP)
+	public boolean outboundMappingEnabled;
+
+	@NotNull
+	@JsonSetter(nulls = Nulls.SKIP)
+	public Integer inboundExternalIdCacheSize;
+
+	@NotNull
+	@JsonSetter(nulls = Nulls.SKIP)
+	public Integer inboundExternalIdCacheRetention;
 }
