@@ -29,30 +29,33 @@ import lombok.Setter;
 @Getter
 @Setter
 public class TopicConfig {
-    private String tenant;
-    private String bootstrapServers;
-    private String topic;
-    private String username;
-    private String password;
-    private String groupId;
-    private Properties defaultPropertiesConsumer;
+	private String tenant;
+	private String bootstrapServers;
+	private String topic;
+	private String username;
+	private String password;
+	private String saslMechanism;
+	private String groupId;
+	private Properties defaultPropertiesConsumer;
 
-    public TopicConfig(String tenant, String bootstrapServers, String topic, String username, String password,
-            String groupId,
-            Properties defaultPropertiesConsumer) {
-        this.tenant = tenant;
-        this.bootstrapServers = bootstrapServers;
-        this.topic = topic;
-        this.username = username;
-        this.password = password;
-        this.groupId = groupId;
-        this.defaultPropertiesConsumer = defaultPropertiesConsumer;
-    }
+	public TopicConfig(String tenant, String bootstrapServers, String topic, String username, String password,
+			String saslMechanism,
+			String groupId,
+			Properties defaultPropertiesConsumer) {
+		this.tenant = tenant;
+		this.bootstrapServers = bootstrapServers;
+		this.topic = topic;
+		this.username = username;
+		this.password = password;
+		this.saslMechanism = saslMechanism;
+		this.groupId = groupId;
+		this.defaultPropertiesConsumer = defaultPropertiesConsumer;
+	}
 
-    @Override
-    public String toString() {
-        return getClass().getSimpleName() +
-                " bootstrapServers='" + bootstrapServers + '\'' +
-                ", topic='" + topic + '\'';
-    }
+	@Override
+	public String toString() {
+		return getClass().getSimpleName() +
+				" bootstrapServers='" + bootstrapServers + '\'' +
+				", topic='" + topic + '\'';
+	}
 }
