@@ -22,16 +22,17 @@
 package dynamic.mapping.core;
 
 public enum Operation {
-    ACTIVATE_MAPPING,
-    CONNECT,
-    DISCONNECT,
-    REFRESH_STATUS_MAPPING,
-    RELOAD_EXTENSIONS,
-    RELOAD_MAPPINGS,
-    RESET_STATUS_MAPPING,
-    REFRESH_NOTIFICATIONS_SUBSCRIPTIONS,
-    DEBUG_MAPPING,
-    SNOOP_MAPPING,
+	ACTIVATE_MAPPING,
+	CONNECT,
+	DISCONNECT,
+	REFRESH_STATUS_MAPPING,
+	RELOAD_EXTENSIONS,
+	RELOAD_MAPPINGS,
+	RESET_STATUS_MAPPING,
+	REFRESH_NOTIFICATIONS_SUBSCRIPTIONS,
+	DEBUG_MAPPING,
+	SNOOP_MAPPING,
 	SNOOP_RESET,
-	RESET_DEPLOYMENT_MAP
+	RESET_DEPLOYMENT_MAP,
+	CLEAR_CACHE
 }
