@@ -192,6 +192,8 @@ public class AsynchronousDispatcherOutbound implements NotificationCallback {
 		@Override
 		public List<ProcessingContext<?>> call() throws Exception {
 			long startTime = System.nanoTime();
+
+			Timer.Sample timer = Timer.start(Metrics.globalRegistry);
 			String tenant = c8yMessage.getTenant();
 			boolean sendPayload = c8yMessage.isSendPayload();
 
@@ -276,12 +278,11 @@ public class AsynchronousDispatcherOutbound implements NotificationCallback {
 										.description("Total number of outbound messages")
 										.tag("connector", processor.connectorClient.getConnectorIdent())
 										.register(Metrics.globalRegistry).increment();
-								Timer.builder("dynmapper_outbound_processing_time")
+								timer.stop(Timer.builder("dynmapper_outbound_processing_time")
 										.tag("tenant", c8yMessage.getTenant())
 										.tag("connector", processor.connectorClient.getConnectorIdent())
 										.description("Processing time of outbound messages")
-										.register(Metrics.globalRegistry)
-										.record(System.nanoTime() - startTime, TimeUnit.NANOSECONDS);
+										.register(Metrics.globalRegistry));
 
 								List<C8YRequest> resultRequests = context.getRequests();
 								if (context.hasError() || resultRequests.stream().anyMatch(r -> r.hasError())) {
