@@ -14,5 +14,5 @@ if (__MODE__ === 'production') {
 export function bootstrap() {
   return platformBrowserDynamic()
     .bootstrapModule(AppModule)
-    .catch(err => console.error(err));
+    .catch((err) => console.error(err));
 }
