@@ -19,13 +19,14 @@
  * @authors Christof Strack
  */
 import { Component, OnInit } from '@angular/core';
-import { FormControl, FormGroup } from '@angular/forms';
+import { FormBuilder, FormControl, FormGroup } from '@angular/forms';
 import { AlertService, gettext } from '@c8y/ngx-components';
 import { BsModalService } from 'ngx-bootstrap/modal';
 import packageJson from '../../package.json';
 import { Feature, Operation, SharedService } from '../shared';
 import { ServiceConfiguration } from './shared/configuration.model';
 import { ConnectorConfigurationService } from '../connector';
+import { HttpStatusCode } from '@angular/common/http';
 
 @Component({
   selector: 'd11r-mapping-service-configuration',
@@ -45,19 +46,23 @@ export class ServiceConfigurationComponent implements OnInit {
     sendMappingStatus: true,
     sendSubscriptionEvents: false,
     sendNotificationLifecycle: false,
-    outboundMappingEnabled: true
+    outboundMappingEnabled: true,
+    inboundExternalIdCacheSize: 0,
+    inboundExternalIdCacheRetention: 0
   };
+  editable2updated: boolean = false;
 
   constructor(
     public bsModalService: BsModalService,
     public alertService: AlertService,
     private sharedService: SharedService,
-	public connectorConfigurationService: ConnectorConfigurationService
+    public connectorConfigurationService: ConnectorConfigurationService,
+    private fb: FormBuilder
   ) {}
 
   ngOnInit() {
     // console.log('Running version', this.version);
-    this.serviceForm = new FormGroup({
+    this.serviceForm = this.fb.group({
       logPayload: new FormControl(''),
       logSubstitution: new FormControl(''),
       logConnectorErrorInBackend: new FormControl(''),
@@ -65,7 +70,9 @@ export class ServiceConfigurationComponent implements OnInit {
       sendMappingStatus: new FormControl(''),
       sendSubscriptionEvents: new FormControl(''),
       sendNotificationLifecycle: new FormControl(''),
-      outboundMappingEnabled: new FormControl('')
+      outboundMappingEnabled: new FormControl(''),
+      inboundExternalIdCacheSize: new FormControl(''),
+      inboundExternalIdCacheRetention: new FormControl('')
     });
 
     this.loadData();
@@ -74,6 +81,22 @@ export class ServiceConfigurationComponent implements OnInit {
   async loadData(): Promise<void> {
     this.serviceConfiguration =
       await this.sharedService.getServiceConfiguration();
+    this.serviceForm.patchValue({
+      logPayload: this.serviceConfiguration.logPayload,
+      logSubstitution: this.serviceConfiguration.logSubstitution,
+      logConnectorErrorInBackend:
+        this.serviceConfiguration.logConnectorErrorInBackend,
+      sendConnectorLifecycle: this.serviceConfiguration.sendConnectorLifecycle,
+      sendMappingStatus: this.serviceConfiguration.sendMappingStatus,
+      sendSubscriptionEvents: this.serviceConfiguration.sendSubscriptionEvents,
+      sendNotificationLifecycle:
+        this.serviceConfiguration.sendNotificationLifecycle,
+      outboundMappingEnabled: this.serviceConfiguration.outboundMappingEnabled,
+      inboundExternalIdCacheSize:
+        this.serviceConfiguration.inboundExternalIdCacheSize,
+      inboundExternalIdCacheRetention:
+        this.serviceConfiguration.inboundExternalIdCacheRetention
+    });
   }
 
   async clickedReconnect2NotificationEndpoint() {
@@ -81,10 +104,22 @@ export class ServiceConfigurationComponent implements OnInit {
       Operation.REFRESH_NOTIFICATIONS_SUBSCRIPTIONS
     );
     // console.log('Details reconnect2NotificationEndpoint', response1);
-    if (response1.status === 201) {
+    if (response1.status === HttpStatusCode.Created) {
       this.alertService.success(gettext('Reconnected successfully.'));
     } else {
       this.alertService.danger(gettext('Failed to reconnect!'));
+    }
+  }
+
+  async clickedClearInboundExternalIdCache() {
+    const response1 = await this.sharedService.runOperation(
+      Operation.CLEAR_CACHE,
+      { cacheId: 'INBOUND_ID_CACHE' }
+    );
+    if (response1.status === HttpStatusCode.Created) {
+      this.alertService.success(gettext('Cache cleared.'));
+    } else {
+      this.alertService.danger(gettext('Failed to clear cache!'));
     }
   }
 
@@ -93,7 +128,7 @@ export class ServiceConfigurationComponent implements OnInit {
       Operation.RESET_DEPLOYMENT_MAP
     );
     // console.log('Details reconnect2NotificationEndpoint', response1);
-    if (response1.status === 201) {
+    if (response1.status === HttpStatusCode.Created) {
       this.alertService.success(gettext('Reset deploymentMap.'));
     } else {
       this.alertService.danger(gettext('Failed to reset deploymentMap!'));
@@ -101,9 +136,7 @@ export class ServiceConfigurationComponent implements OnInit {
   }
 
   async clickedSaveServiceConfiguration() {
-    const conf: ServiceConfiguration = {
-      ...this.serviceConfiguration
-    };
+    const conf = this.serviceForm.value;
     const response = await this.sharedService.updateServiceConfiguration(conf);
     if (response.status < 300) {
       this.alertService.success(gettext('Update successful'));
