@@ -20,11 +20,33 @@
  */
 import { Component } from '@angular/core';
 import { CellRendererContext } from '@c8y/ngx-components';
+import { MappingService } from '../core/mapping.service';
 
+/**
+ * The example component for custom cell renderer.
+ * It gets `context` with the current row item and the column.
+ * Additionally, a service is injected to provide a helper method.
+ * The template displays the icon and the label with additional styling.
+ */
 @Component({
-  selector: 'd11r-mapping-renderer-snooped',
-  template: '<span>{{ context.value.snoopedTemplates ? context.value.snoopedTemplates.length: "" }}</span>'
+  template: `
+    <button
+      class="btn btn-link"
+      title="{{ context.item.id }}"
+      (click)="updateMapping()"
+      style="padding-top: 0px; padding-bottom: 10px;"
+    >
+      {{ context.value }}
+    </button>
+  `
 })
-export class SnoopedTemplateRendererComponent {
-  constructor(public context: CellRendererContext) {}
+export class MappingIdCellRendererComponent {
+  constructor(
+    public context: CellRendererContext,
+    public mappingService: MappingService
+  ) {}
+  async updateMapping() {
+    const { mapping } = this.context.item;
+    this.mappingService.initiateUpdateMapping({ mapping, id: mapping.id });
+  }
 }
