@@ -41,6 +41,7 @@ import { ExtensionService } from '../../extension';
 import {
   API,
   COLOR_HIGHLIGHTED,
+  ConnectorType,
   DeploymentMapEntry,
   Direction,
   Extension,
@@ -56,7 +57,7 @@ import {
 } from '../../shared';
 import { JsonEditor2Component } from '../../shared/editor2/jsoneditor2.component';
 import { MappingService } from '../core/mapping.service';
-import { EditSubstitutionComponent } from '../edit/edit-substitution-modal.component';
+import { EditSubstitutionComponent } from '../substitution/edit/edit-substitution-modal.component';
 import { C8YRequest } from '../processor/processor.model';
 import { ValidationError } from '../shared/mapping.model';
 import { EditorMode } from '../shared/stepper-model';
@@ -84,6 +85,7 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
   @Output() cancel = new EventEmitter<any>();
   @Output() commit = new EventEmitter<Mapping>();
   private _deploymentMapEntry: DeploymentMapEntry;
+  isButtonDisabled$: BehaviorSubject<boolean> = new BehaviorSubject(true);
 
   @Input()
   get deploymentMapEntry(): DeploymentMapEntry {
@@ -91,9 +93,8 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
   }
   set deploymentMapEntry(value: DeploymentMapEntry) {
     this._deploymentMapEntry = value;
-    this.deploymentMapEntryChange.emit(value);
+    // console.log('New setDeploymentMap', this._deploymentMapEntry);
   }
-  @Output() deploymentMapEntryChange = new EventEmitter<any>();
 
   ValidationError = ValidationError;
   Direction = Direction;
@@ -137,8 +138,10 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
   sourceSystem: string;
   targetSystem: string;
 
-  editorOptionsSource: any = {};
-  editorOptionsTarget: any = {};
+  editorOptionsSourceStep3: any = {};
+  editorOptionsSourceStep4: any = {};
+  editorOptionsTargetStep3: any = {};
+  editorOptionsTargetStep4: any = {};
   editorOptionsTesting: any = {};
 
   selectedSubstitution: number = -1;
@@ -148,10 +151,14 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
   expertMode: boolean = false;
   templatesInitialized: boolean = false;
 
-  @ViewChild('editorSource', { static: false })
-  editorSource: JsonEditor2Component;
-  @ViewChild('editorTarget', { static: false })
-  editorTarget: JsonEditor2Component;
+  @ViewChild('editorSourceStep4', { static: false })
+  editorSourceStep4: JsonEditor2Component;
+  @ViewChild('editorTargetStep4', { static: false })
+  editorTargetStep4: JsonEditor2Component;
+  @ViewChild('editorSourceStep3', { static: false })
+  editorSourceStep3: JsonEditor2Component;
+  @ViewChild('editorTargetStep3', { static: false })
+  editorTargetStep3: JsonEditor2Component;
   editorTestingResponse: JsonEditor2Component;
   @ViewChild(SubstitutionRendererComponent, { static: false })
   substitutionChild: SubstitutionRendererComponent;
@@ -159,6 +166,7 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
   extensions: Map<string, Extension> = new Map();
   extensionEvents$: BehaviorSubject<string[]> = new BehaviorSubject([]);
   onDestroy$ = new Subject<void>();
+  supportsMessageContext: boolean;
 
   constructor(
     public bsModalService: BsModalService,
@@ -167,6 +175,25 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
     private alertService: AlertService,
     private elementRef: ElementRef
   ) {}
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  deploymentMapEntryChange(e) {
+    // console.log(
+    //   'New getDeploymentMap',
+    //   this._deploymentMapEntry,
+    //   !this._deploymentMapEntry?.connectors
+    // );
+    this.isButtonDisabled$.next(
+      !this._deploymentMapEntry?.connectors ||
+        this._deploymentMapEntry?.connectors?.length == 0
+    );
+    this.supportsMessageContext =
+      this._deploymentMapEntry.connectorsDetailed?.some(
+        (con) => con.connectorType == ConnectorType.KAFKA
+      );
+
+    // console.log('New setDeploymentMap from grid', e);
+  }
 
   ngOnInit() {
     // console.log('mapping-stepper', this._deploymentMapEntry, this.deploymentMapEntry);
@@ -335,55 +362,12 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
             }
           }
         ]
-      },
-
-    //   {
-    //     fieldGroup: [
-    //       {
-    //         className: 'col-lg-5 col-lg-offset-1 text-monospace font-smaller',
-    //         key: 'sourceExpression.result',
-    //         type: 'textarea-custom',
-    //         wrappers: ['custom-form-wrapper'],
-    //         templateOptions: {
-    //           class: 'no-resize',
-    //           disabled: false,
-    //           readonly: false,
-    //           customWrapperClass: 'm-b-4'
-    //         },
-    //         expressionProperties: {
-    //           'templateOptions.label': () =>
-    //             `Source Result [${this.substitutionModel.sourceExpression?.resultType}]`,
-    //           'templateOptions.value': () => {
-    //             return `${this.substitutionModel.sourceExpression?.result}`;
-    //           }
-    //         }
-    //       },
-    //       {
-    //         className: 'col-lg-5 text-monospace font-smaller',
-    //         key: 'targetExpression.result',
-    //         type: 'textarea-custom',
-    //         wrappers: ['custom-form-wrapper'],
-    //         templateOptions: {
-    //           class: 'input',
-    //           disabled: true,
-    //           readonly: true,
-    //           customWrapperClass: 'm-b-4'
-    //         },
-    //         expressionProperties: {
-    //           'templateOptions.label': () =>
-    //             `Target Result [${this.substitutionModel.targetExpression?.resultType}]`,
-    //           'templateOptions.value': () => {
-    //             return `${this.substitutionModel.targetExpression?.result}`;
-    //           }
-    //         }
-    //       }
-    //     ]
-    //   }
+      }
     ];
 
     this.setTemplateForm();
-    this.editorOptionsSource = {
-      ...this.editorOptionsSource,
+    this.editorOptionsSourceStep3 = {
+      ...this.editorOptionsSourceStep3,
       mode: 'tree',
       mainMenuBar: true,
       navigationBar: false,
@@ -392,11 +376,30 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
       name: 'message'
     };
 
-    this.editorOptionsTarget = {
-      ...this.editorOptionsTarget,
+    this.editorOptionsTargetStep3 = {
+      ...this.editorOptionsTargetStep3,
       mode: 'tree',
       mainMenuBar: true,
       navigationBar: false,
+      statusBar: true
+    };
+
+    this.editorOptionsSourceStep4 = {
+      ...this.editorOptionsSourceStep4,
+      mode: 'tree',
+      mainMenuBar: true,
+      navigationBar: false,
+      statusBar: false,
+      readOnly: true,
+      name: 'message'
+    };
+
+    this.editorOptionsTargetStep4 = {
+      ...this.editorOptionsTargetStep4,
+      mode: 'tree',
+      mainMenuBar: true,
+      navigationBar: false,
+      readOnly: true,
       statusBar: true
     };
 
@@ -465,7 +468,7 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
       this.substitutionFormly.get('pathSource').setErrors(null);
 
       const r: JSON = await this.mappingService.evaluateExpression(
-        this.editorSource?.get(),
+        this.editorSourceStep4?.get(),
         this.substitutionFormly.get('pathSource').value
       );
       this.substitutionModel.sourceExpression = {
@@ -511,24 +514,26 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
     this.substitutionModel.pathTarget = path;
   }
 
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   onTemplateSourceChanged(content: Content) {
-    if (_.has(content, 'text') && content['text']) {
-      this.templateSource = JSON.parse(content['text']);
-      // this.mapping.source = content['text'];‚
-    } else {
-      this.templateSource = content['json'];
-      // this.mapping.source = JSON.stringify(content['json']);
-    }
+    // if (_.has(content, 'text') && content['text']) {
+    //   this.templateSource = JSON.parse(content['text']);
+    //   // this.mapping.source = content['text'];‚
+    // } else {
+    //   this.templateSource = content['json'];
+    //   // this.mapping.source = JSON.stringify(content['json']);
+    // }
   }
 
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   onTemplateTargetChanged(content: Content) {
-    if (_.has(content, 'text') && content['text']) {
-      this.templateTarget = JSON.parse(content['text']);
-      // this.mapping.target = content['text'];
-    } else {
-      this.templateTarget = content['json'];
-      // this.mapping.target = JSON.stringify(content['json']);
-    }
+    // if (_.has(content, 'text') && content['text']) {
+    //   this.templateTarget = JSON.parse(content['text']);
+    //   // this.mapping.target = content['text'];
+    // } else {
+    //   this.templateTarget = content['json'];
+    //   // this.mapping.target = JSON.stringify(content['json']);
+    // }
   }
 
   async updateTargetExpressionResult() {
@@ -540,7 +545,7 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
       };
       this.substitutionFormly.get('pathTarget').setErrors(null);
       const r: JSON = await this.mappingService.evaluateExpression(
-        this.editorTarget?.get(),
+        this.editorTargetStep4?.get(),
         path
       );
       this.substitutionModel.targetExpression = {
@@ -562,7 +567,7 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
         this.substitutionModel.targetExpression.severity = 'text-info';
       } else if (path == '$') {
         this.substitutionModel.targetExpression.msgTxt = `By specifying "$" you selected the root of the target 
-        template and this rersults in merging the source expression with the target template.`;
+        template and this result in merging the source expression with the target template.`;
         this.substitutionModel.targetExpression.severity = 'text-warning';
       }
     } catch (error) {
@@ -582,10 +587,10 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
     return {
       ...this.mapping,
       source: reduceSourceTemplate(
-        this.editorSource ? this.editorSource?.get() : {},
+        this.editorSourceStep4 ? this.editorSourceStep4?.get() : {},
         patched
       ), // remove array "_TOPIC_LEVEL_" since it should not be stored
-      target: reduceTargetTemplate(this.editorTarget?.get(), patched), // remove patched attributes, since it should not be stored
+      target: reduceTargetTemplate(this.editorTargetStep4?.get(), patched), // remove patched attributes, since it should not be stored
       lastUpdate: Date.now()
     };
   }
@@ -610,7 +615,7 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
         levels
       );
     }
-    this.editorTarget.set(this.templateTarget);
+    this.editorTargetStep4.set(this.templateTarget);
   }
 
   async onCancelButton() {
@@ -636,7 +641,40 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
   }): Promise<void> {
     // ('OnNextStep', event.step.label, this.mapping);
     this.step = event.step.label;
-    if (this.step == 'General settings') {
+    if (this.step == 'Add and select connector') {
+      if (
+        this.deploymentMapEntry.connectors &&
+        this.deploymentMapEntry.connectors.length == 0
+      ) {
+        // const initialState = {
+        //   title: 'No connector selected',
+        //   message:
+        //     'To apply the mapping to messages you should select at least one connector, unless you want to change this later! Do you want to continue?',
+        //   labels: {
+        //     ok: 'Continue',
+        //     cancel: 'Close'
+        //   }
+        // };
+        // const confirmContinuingModalRef: BsModalRef = this.bsModalService.show(
+        //   ConfirmationModalComponent,
+        //   { initialState }
+        // );
+        // confirmContinuingModalRef.content.closeSubject.subscribe(
+        //   async (confirmation: boolean) => {
+        //     // console.log('Confirmation result:', confirmation);
+        //     if (confirmation) {
+        //       event.stepper.next();
+        //     }
+        //     confirmContinuingModalRef.hide();
+        //   }
+        // );
+        // this.alertService.warning(
+        //   'To apply the mapping to messages you have to select at least one connector. Go back, unless you only want to assign a connector later!'
+        // );
+      } else {
+        event.stepper.next();
+      }
+    } else if (this.step == 'General settings') {
       this.templateModel.mapping = this.mapping;
       // console.log(
       //  'Populate jsonPath if wildcard:',
@@ -666,12 +704,27 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
       }
       event.stepper.next();
     } else if (this.step == 'Define substitutions') {
+      this.expandTemplates();
       this.getTemplateForm();
-      const testSourceTemplate = this.editorSource
-        ? this.editorSource.get()
-        : {};
-      this.editorTestingPayloadTemplateEmitter.emit(testSourceTemplate);
+      //   const testSourceTemplate = this.editorSourceStep4
+      //     ? this.editorSourceStep4.get()
+      //     : {};
+      // this.editorTestingPayloadTemplateEmitter.emit(testSourceTemplate);
+      this.editorTestingPayloadTemplateEmitter.emit(this.templateSource);
       this.onSelectSubstitution(0);
+      event.stepper.next();
+    } else if (this.step == 'Select templates') {
+      this.templateSource = this.editorSourceStep3?.get();
+      this.templateTarget = this.editorTargetStep3?.get();
+      console.log(
+        'onNextStep before',
+        event.step.label,
+        this.mapping,
+        this.editorSourceStep3?.get(),
+        this.getCurrentMapping(true),
+        this.templateSource,
+        this.templateTarget
+      );
       event.stepper.next();
     } else {
       event.stepper.next();
@@ -682,7 +735,15 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
     stepper: C8yStepper;
     step: CdkStep;
   }): Promise<void> {
-    // console.log('onBackStep', event.step.label, this.mapping);
+    // console.log(
+    //   'onBackStep before',
+    //   event.step.label,
+    //   this.mapping,
+    //   this.getCurrentMapping(false),
+    //   this.getCurrentMapping(true),
+    //   this.templateSource,
+    //   this.templateTarget
+    // );
     this.step = event.step.label;
     if (this.step == 'Test mapping') {
       const editorTestingRequestRef =
@@ -691,10 +752,24 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
         editorTestingRequestRef.setAttribute('schema', undefined);
       }
     } else if (this.step == 'Select templates') {
-      this.mapping = this.getCurrentMapping(false);
+      // this.mapping = this.getCurrentMapping(true);
+      // this.expandTemplates();
     } else if (this.step == 'Define substitutions') {
-      this.mapping = this.getCurrentMapping(false);
+      // this.mapping = this.getCurrentMapping(true);
+      // this.expandTemplates();
     }
+
+    this.mapping = this.getCurrentMapping(true);
+    this.expandTemplates();
+    // console.log(
+    //   'onBackStep after',
+    //   event.step.label,
+    //   this.mapping,
+    //   this.getCurrentMapping(false),
+    //   this.getCurrentMapping(true),
+    //   this.templateSource,
+    //   this.templateTarget
+    // );
     event.stepper.previous();
   }
 
@@ -821,8 +896,20 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
     this.mapping.snoopStatus = SnoopStatus.STOPPED;
   }
 
-  async onTargetTemplateChanged(templateTarget) {
-    this.templateTarget = templateTarget;
+  async onTargetAPIChanged(changedTargetAPI) {
+    if (this.stepperConfiguration.direction == Direction.INBOUND) {
+      this.mapping.target = SAMPLE_TEMPLATES_C8Y[changedTargetAPI];
+      this.mapping.source = getExternalTemplate(this.mapping);
+      this.schemaUpdateTarget.emit(
+        getSchema(this.mapping.targetAPI, this.mapping.direction, true)
+      );
+    } else {
+      this.mapping.source = SAMPLE_TEMPLATES_C8Y[changedTargetAPI];
+      this.mapping.target = getExternalTemplate(this.mapping);
+      this.schemaUpdateSource.emit(
+        getSchema(this.mapping.targetAPI, this.mapping.direction, false)
+      );
+    }
   }
 
   async updateTestResult(result) {
@@ -867,7 +954,8 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
       const initialState = {
         substitution: _.clone(this.mapping.substitutions[selected]),
         mapping: this.mapping,
-        stepperConfiguration: this.stepperConfiguration
+        stepperConfiguration: this.stepperConfiguration,
+        isUpdate: true
       };
       if (
         this.substitutionModel.sourceExpression?.severity != 'text-danger' &&
@@ -902,7 +990,7 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
     this.mapping.substitutions.forEach((s, index) => {
       if (sub.pathTarget == s.pathTarget) {
         duplicateSubstitutionIndex = index;
-		duplicate = this.mapping.substitutions[index];
+        duplicate = this.mapping.substitutions[index];
       }
     });
     const isDuplicate = duplicateSubstitutionIndex != -1;
@@ -939,17 +1027,17 @@ export class MappingStepperComponent implements OnInit, OnDestroy {
       this.selectedSubstitution = selected;
       this.substitutionModel = _.clone(this.mapping.substitutions[selected]);
       this.substitutionModel.stepperConfiguration = this.stepperConfiguration;
-      await this.editorSource?.setSelectionToPath(
+      await this.editorSourceStep4?.setSelectionToPath(
         this.substitutionModel.pathSource
       );
-      await this.editorTarget.setSelectionToPath(
+      await this.editorTargetStep4.setSelectionToPath(
         this.substitutionModel.pathTarget
       );
     }
   }
 
   onTemplateChanged(templateTarget: any): void {
-    this.editorTarget.set(templateTarget);
+    this.editorTargetStep4.set(templateTarget);
   }
 
   ngOnDestroy() {
