@@ -24,12 +24,8 @@ import {
   Input,
   OnInit,
   Output,
-  OnDestroy,
   ViewChild,
-  AfterViewInit,
-  ElementRef,
-  Renderer2,
-  AfterViewChecked
+  AfterViewInit
 } from '@angular/core';
 import {
   ActionControl,
@@ -43,10 +39,11 @@ import {
 import { BsModalRef, BsModalService } from 'ngx-bootstrap/modal';
 import {
   BehaviorSubject,
+  combineLatest,
   from,
   Observable,
-  ReplaySubject,
-  Subject
+  Subject,
+  take
 } from 'rxjs';
 
 import * as _ from 'lodash';
@@ -58,7 +55,7 @@ import {
 } from '../connector-log/connector-status.model';
 import { DeploymentMapEntry } from '../model/shared.model';
 import { uuidCustom } from '../model/util';
-import { ConfigurationConfigurationModalComponent } from './connector-configuration-modal.component';
+import { ConfigurationConfigurationModalComponent } from './create/connector-configuration-modal.component';
 import {
   ConnectorConfiguration,
   ConnectorSpecification
@@ -83,16 +80,15 @@ export class ConnectorConfigurationComponent implements OnInit, AfterViewInit {
   }
   set deploymentMapEntry(value: DeploymentMapEntry) {
     this._deploymentMapEntry = value;
-    this.deploymentMapEntryChange.emit(value);
   }
   @Output() deploymentMapEntryChange = new EventEmitter<any>();
   selected: string[] = [];
-  selected$: Subject<string[]> = new BehaviorSubject([]);
+  selected$: Subject<string[]>;
   selectedAll: boolean = false;
   monitoring$: Observable<ConnectorStatus>;
   specifications: ConnectorSpecification[] = [];
   configurations: ConnectorConfiguration[];
-  configurations$: Subject<ConnectorConfiguration[]> = new ReplaySubject(1);
+  configurations$: Observable<ConnectorConfiguration[]>;
   StatusEventTypes = StatusEventTypes;
   pagination: Pagination = {
     pageSize: 30,
@@ -107,9 +103,7 @@ export class ConnectorConfigurationComponent implements OnInit, AfterViewInit {
   constructor(
     private bsModalService: BsModalService,
     private connectorConfigurationService: ConnectorConfigurationService,
-    private alertService: AlertService,
-    private el: ElementRef,
-    private renderer: Renderer2
+    private alertService: AlertService
   ) {}
 
   ngAfterViewInit(): void {
@@ -175,7 +169,8 @@ export class ConnectorConfigurationComponent implements OnInit, AfterViewInit {
         path: 'name',
         filterable: false,
         sortOrder: 'asc',
-        visible: true
+        visible: true,
+        gridTrackSize: '30%'
       },
       {
         name: 'connectorType',
@@ -184,7 +179,7 @@ export class ConnectorConfigurationComponent implements OnInit, AfterViewInit {
         filterable: false,
         sortOrder: 'asc',
         visible: true,
-        gridTrackSize: '20%'
+        gridTrackSize: '30%'
       },
       {
         header: 'Status',
@@ -193,7 +188,7 @@ export class ConnectorConfigurationComponent implements OnInit, AfterViewInit {
         filterable: false,
         sortable: true,
         cellRendererComponent: ConnectorStatusRendererComponent,
-        gridTrackSize: '20%'
+        gridTrackSize: '10%'
       },
       {
         header: 'Enabled',
@@ -202,37 +197,38 @@ export class ConnectorConfigurationComponent implements OnInit, AfterViewInit {
         filterable: false,
         sortable: true,
         cellRendererComponent: StatusEnabledRendererComponent,
-        gridTrackSize: '15%'
+        gridTrackSize: '10%'
       }
     );
+
+    this.configurations$ =
+      this.connectorConfigurationService.getConnectorConfigurationsWithLiveStatus();
+
     this.selected = this.deploymentMapEntry?.connectors ?? [];
+    this.selected$ = new BehaviorSubject(this.selected);
 
-    this.selected$.next(this.selected);
-    this.selected$.subscribe((se) => {
-      if (this.selectable) {
-        this.deploymentMapEntry.connectors = se;
+    combineLatest([this.selected$, this.configurations$]).subscribe(
+      ([se, conf]) => {
+        this.configurations = conf;
+        if (this.selectable) {
+          this.deploymentMapEntry.connectors = se;
+          this.deploymentMapEntry.connectorsDetailed = conf.filter((con) =>
+            se.includes(con.ident)
+          );
+          this.deploymentMapEntryChange.emit(this.deploymentMapEntry);
+          if (this.readOnly)
+            this.configurations?.forEach(
+              (conf) => (conf['checked'] = this.selected.includes(conf.ident))
+            );
+        }
       }
-    });
+    );
 
-    from(
-      this.connectorConfigurationService.getConnectorSpecifications()
-    ).subscribe((specs) => {
-      this.specifications = specs;
-    });
-
-    this.connectorConfigurationService
-      .getRealtimeConnectorConfigurations()
-      .subscribe((confs) => {
-        this.configurations$.next(confs);
+    from(this.connectorConfigurationService.getConnectorSpecifications())
+      .pipe(take(1))
+      .subscribe((specs) => {
+        this.specifications = specs;
       });
-
-    this.configurations$.subscribe((confs) => {
-      this.configurations = confs;
-      if (this.selectable && this.readOnly)
-        this.configurations?.forEach(
-          (conf) => (conf['checked'] = this.selected.includes(conf.ident))
-        );
-    });
   }
 
   public onSelectToggle(id: string) {
@@ -260,7 +256,7 @@ export class ConnectorConfigurationComponent implements OnInit, AfterViewInit {
   }
   public onSelectionChanged(selected: any) {
     this.selected = selected;
-    this.selected$.next(this.selected);
+    this.selected$?.next(this.selected);
   }
 
   public isSelectedAll(): boolean {
