@@ -23,21 +23,22 @@ import { CellRendererContext } from '@c8y/ngx-components';
 
 @Component({
   selector: 'd11r-mapping-renderer-qos',
-  template: `<div class="c8y-realtime p-l-8">
-    <span
-      class="c8y-pulse animated pulse animation-slow"
-      [ngClass]="{
-        active: (context.item?.status$ | async) === 'CONNECTED',
-        inactive: (context.item?.status$ | async) !== 'CONNECTED'
-      }"
-    ></span
+  template: `<span
+    [class]="
+      'text-12 tag ' +
+      ((context.item?.status$ | async) === 'CONNECTED'
+        ? 'tag--success'
+        : (context.item?.status$ | async) === 'FAILED'
+          ? 'tag--danger'
+          : 'tag--default')
+    "
     >{{ context.item?.status$ | async }}
-  </div>`
+  </span> `
 })
 export class ConnectorStatusRendererComponent {
   constructor(public context: CellRendererContext) {
-	// console.log(`Connector: ${context.item.name}`, context.item);
-	// const status$ = context.item?.status$;
-	// status$.subscribe( n => console.log(`Connector new status: ${n}`));
+    // console.log(   `Connector: ${context.item.name}`, context.item?.status$ | async);
+    // const status$ = context.item?.status$;
+    // status$.subscribe( n => console.log(`Connector new status: ${n}`));
   }
 }
