@@ -18,72 +18,76 @@
  *
  * @authors Christof Strack
  */
-import { Injectable } from '@angular/core';
-import { EventService, FetchClient, IEvent, Realtime } from '@c8y/client';
+import { inject, Injectable } from '@angular/core';
+import { EventService, FetchClient } from '@c8y/client';
 import {
-  BASE_URL,
   CONNECTOR_FRAGMENT,
   ConnectorStatus,
-  PATH_STATUS_CONNECTORS_ENDPOINT,
-  SharedService
+  ConnectorStatusEvent,
+  SharedService,
+  StatusEventTypes
 } from '../shared';
 
-import { merge, Observable, ReplaySubject, Subject } from 'rxjs';
+import { BehaviorSubject, from, merge, Observable } from 'rxjs';
+import { filter, map, scan, shareReplay, switchMap, tap } from 'rxjs/operators';
 import {
-  filter,
-  map,
-  scan,
-  share,
-  shareReplay,
-  switchMap,
-  tap
-} from 'rxjs/operators';
+  EventRealtimeService,
+  RealtimeSubjectService
+} from '@c8y/ngx-components';
 
-@Injectable({ providedIn: 'root' })
+@Injectable({
+  providedIn: 'root'
+})
 export class ConnectorStatusService {
   constructor(
     private client: FetchClient,
     private eventService: EventService,
     private sharedService: SharedService
   ) {
-    this.realtime = new Realtime(this.client);
+    this.eventRealtimeService = new EventRealtimeService(
+      inject(RealtimeSubjectService)
+    );
     this.startConnectorStatusLogs();
   }
+  private eventRealtimeService: EventRealtimeService;
 
   private _agentId: string;
-  private initialized: boolean = false;
-  private realtime: Realtime;
-  private subscriptionEvents: any;
-  private filterStatusLog = {
-    eventType: 'ALL',
-    // eventType: StatusEventTypes.STATUS_CONNECTOR_EVENT_TYPE,
-    connectorIdent: 'ALL'
-  };
-  private triggerLogs$: Subject<any> = new Subject();
-  private realtimeConnectorStatus$: Subject<IEvent> = new Subject();
-  private statusLogs$: Subject<any[]> = new ReplaySubject(1);
+  private readonly ALL: string = 'ALL';
 
-  getStatusLogs(): Observable<any[]> {
+  private readonly RESET = {
+    connectorIdent: this.ALL,
+    connectorName: 'EMPTY',
+    status: ConnectorStatus.UNKNOWN,
+    type: StatusEventTypes.ALL,
+    message: '_RESET_'
+  };
+
+  private filterStatusLog = this.RESET;
+
+  private triggerLogs$: BehaviorSubject<ConnectorStatusEvent[]> =
+    new BehaviorSubject([this.RESET]);
+
+  private statusLogs$: Observable<ConnectorStatusEvent[]>;
+
+  getStatusLogs(): Observable<ConnectorStatusEvent[]> {
+    // console.log('Calling: getStatusLogs');
     return this.statusLogs$;
   }
 
   async startConnectorStatusLogs() {
-    console.log('Calling: startConnectorStatusLogs', this.initialized);
-    if (!this.initialized) {
-      this.startConnectorStatusSubscriptions();
+    // console.log('Calling: startConnectorStatusLogs');
+    if (!this.statusLogs$) {
       await this.initConnectorLogsRealtime();
-      this.initialized = true;
     }
-    this.triggerLogs$.next([{ type: 'reset' }]);
   }
 
   updateStatusLogs(filter: any) {
-    this.triggerLogs$.next([{ type: 'reset' }]);
-    this.filterStatusLog = filter;
-  }
-
-  async stopConnectorStatusLogs() {
-    this.realtime.unsubscribe(this.subscriptionEvents);
+    const updatedFilter = {
+      ...this.RESET,
+      ...filter
+    };
+    this.filterStatusLog = updatedFilter;
+    this.triggerLogs$.next([updatedFilter]);
   }
 
   async initConnectorLogsRealtime() {
@@ -96,14 +100,14 @@ export class ConnectorStatusService {
     // );
     const filteredConnectorStatus$ = this.triggerLogs$.pipe(
       // tap((x) => console.log('TriggerLogs In', x)),
-      switchMap(() => {
+      switchMap((x) => {
         const filter = {
           pageSize: 5,
           withTotalPages: false,
           source: this._agentId
         };
-        if (this.filterStatusLog.eventType !== 'ALL') {
-          filter['type'] = this.filterStatusLog.eventType;
+        if (x[0]?.type !== this.ALL) {
+          filter['type'] = x[0]?.type;
         }
         return this.eventService.list(filter);
       }),
@@ -118,7 +122,7 @@ export class ConnectorStatusService {
       ),
       map((events) =>
         events.filter((event) => {
-          return this.filterStatusLog.connectorIdent == 'ALL'
+          return this.filterStatusLog.connectorIdent == this.ALL
             ? true
             : event.connectorIdent == this.filterStatusLog.connectorIdent;
         })
@@ -126,77 +130,54 @@ export class ConnectorStatusService {
       //   tap((x) => console.log('TriggerLogs Out', x))
     );
 
-    const realtimeConnectorStatusRealtime$ = this.realtimeConnectorStatus$.pipe(
-      // tap((x) => console.log('IncomingRealtime In', x)),
-      filter((event) => {
-        return (
-          Object.keys(event).length !== 0 &&
-          (this.filterStatusLog.eventType == 'ALL'
-            ? true
-            : event.type == this.filterStatusLog.eventType) &&
-          (this.filterStatusLog.connectorIdent == 'ALL'
-            ? true
-            : event[CONNECTOR_FRAGMENT]?.connectorIdent ==
-              this.filterStatusLog.connectorIdent)
-        );
-      }),
-      map((event) => {
-        event[CONNECTOR_FRAGMENT].type = event?.type;
-        return [event[CONNECTOR_FRAGMENT]];
-      })
-    );
-
     //  const refreshedConnectorStatus$: Observable<any> =
-    merge(
+    this.statusLogs$ = merge(
       filteredConnectorStatus$,
-      realtimeConnectorStatusRealtime$,
+      this.getAllConnectorStatusEvents(),
       this.triggerLogs$
-    )
-      .pipe(
-        // tap((i) => console.log('Items', i)),
-        scan((acc, val) => {
-          let sortedAcc;
-          if (val[0]?.type == 'reset') {
-            // console.log('Reset loaded logs!');
-            sortedAcc = [];
-          } else {
-            sortedAcc = val.concat(acc);
-          }
-          sortedAcc = sortedAcc.slice(0, 9);
-          return sortedAcc;
-        }, []),
-        tap((logs) => this.statusLogs$.next(logs))
-        // shareReplay(1)
-      )
-      .subscribe();
+    ).pipe(
+      // tap((i) => console.log('Items', i)),
+      scan((acc, val) => {
+        let sortedAcc;
+        if (val[0]?.message == '_RESET_') {
+          // console.log('Reset loaded logs!');
+          sortedAcc = [];
+        } else {
+          sortedAcc = val.concat(acc);
+        }
+        sortedAcc = sortedAcc.slice(0, 9);
+        return sortedAcc;
+      }, []),
+      shareReplay(1)
+    );
   }
 
-  async startConnectorStatusSubscriptions(): Promise<void> {
-    if (!this._agentId) {
-      this._agentId = await this.sharedService.getDynamicMappingServiceAgent();
-    }
+  private getAllConnectorStatusEvents(): Observable<ConnectorStatusEvent[]> {
     // console.log('Started subscriptions:', this._agentId);
 
     // subscribe to event stream
-    this.subscriptionEvents = this.realtime.subscribe(
-      `/events/${this._agentId}`,
-      this.updateRealtimeConnectorStatus
+    this.eventRealtimeService.start();
+    return from(this.sharedService.getDynamicMappingServiceAgent()).pipe(
+      switchMap((agentId) => {
+        return this.eventRealtimeService.onAll$(agentId);
+      }),
+      map((p) => p['data']),
+      map((e) => {
+        e[CONNECTOR_FRAGMENT].type = e['type'];
+        return e[CONNECTOR_FRAGMENT];
+      }),
+      filter((e) =>
+        this.filterStatusLog.type == this.ALL
+          ? true
+          : e['type'] == this.filterStatusLog.type
+      ),
+      filter((e) =>
+        this.filterStatusLog.connectorIdent == this.ALL
+          ? true
+          : e.connectorIdent == this.filterStatusLog.connectorIdent
+      ),
+      map((e) => [e]),
+      tap((l) => console.log('StatusLogs:', l))
     );
-  }
-
-  private updateRealtimeConnectorStatus = async (p: object) => {
-    const payload = p['data']['data'];
-    this.realtimeConnectorStatus$.next(payload);
-  };
-
-  async getConnectorStatus(): Promise<ConnectorStatus> {
-    const response = await this.client.fetch(
-      `${BASE_URL}/${PATH_STATUS_CONNECTORS_ENDPOINT}`,
-      {
-        method: 'GET'
-      }
-    );
-    const result = await response.json();
-    return result;
   }
 }
