@@ -12,5 +12,4 @@ export * from './camel-case.pipe';
 export * from './connector-log/connector-status.model';
 export * from './connector-configuration/connector.model';
 export * from './connector-log/connector-status.component';
-export * from './connector-configuration/connector-configuration-modal.component';
-
+export * from './connector-configuration/create/connector-configuration-modal.component';
